review: address #662 round 2
- Move the five "*** enabled*" Info logs out of buildOAuthServerOptions
  into a new logEnabledOAuthOptions, called only after
  oauth.NewServerWithCombined returns nil error. Constructor failures
  no longer produce false "enabled" log lines.
- Drop the WithSecurityEventRateLimiter bullet from CHANGELOG ### Added;
  it's internal infra, fold the one-line summary into the existing
  ### Changed mcp-oauth bump line.
- Rename oauth_http_trusted_redirect_uris_test.go to
  new_oauth_server_config_test.go: the file scope is the mapper, not
  one field.
- Rename TestSSOLifecycleOptions_Count to
  TestSSOLifecycleOptions_WiresThreeNonNilCallbacks; add a comment
  flagging it's a count pin, not a kind pin.
- Split newOAuthServerConfig + buildOAuthServerOptions out of
  oauth_http.go into oauth_server_config.go. oauth_http.go drops from
  888 to 812 lines; new helpers are isolated and discoverable.
- cmd/test.go: collapse the workflow/mcpserver/service triplet into a
  single validTestConcepts slice consumed by both shell completion
  and the parser; error message iterates the slice so adding a fourth
  concept stays single-source.
